+++ src/test/java/io/appium/java_client/MobileDriverIOSTest.java
@@ -17,9 +17,9 @@
 
 package io.appium.java_client;
 
+import io.appium.java_client.ios.IOSDriver;
 import io.appium.java_client.remote.HideKeyboardStrategy;
 import io.appium.java_client.remote.MobileCapabilityType;
-import io.appium.java_client.remote.MobilePlatform;
 
 import org.junit.After;
 import org.junit.Before;
@@ -38,7 +38,7 @@ import static org.junit.Assert.assertEquals;
  */
 public class MobileDriverIOSTest {
 
-  private AppiumDriver driver;
+  private IOSDriver driver;
 
   @Before
   public void setup() throws Exception {
@@ -47,10 +47,10 @@ public class MobileDriverIOSTest {
     DesiredCapabilities capabilities = new DesiredCapabilities();
     capabilities.setCapability(MobileCapabilityType.BROWSER_NAME, "");
     capabilities.setCapability(MobileCapabilityType.PLATFORM_VERSION, "7.1");
-    capabilities.setCapability(MobileCapabilityType.PLATFORM_NAME, MobilePlatform.IOS);
+    //capabilities.setCapability(MobileCapabilityType.PLATFORM_NAME, MobilePlatform.IOS);
     capabilities.setCapability(MobileCapabilityType.DEVICE_NAME, "iPhone Simulator");
     capabilities.setCapability(MobileCapabilityType.APP, app.getAbsolutePath());
-    driver = new AppiumDriver(new URL("http://127.0.0.1:4723/wd/hub"), capabilities);
+    driver = new IOSDriver(new URL("http://127.0.0.1:4723/wd/hub"), capabilities);
   }
 
   @After
